requirements.txt
# PyTorch and related packages
# Jupyter for notebook execution
jupyter>=1.0.0
numpy>=1.24.4
matplotlib>=3.9.1
pillow>=9.4.0
opencv-python>=4.7.0

# SAM2 package
sam2>=0.1.0

tqdm>=4.66.1
hydra-core>=1.3.2
iopath>=0.1.10
eva-decord>=0.6.1